requirements.txt
torch>=2.0.0
torchdiffeq>=0.2.3
MDAnalysis>=2.4.0
numpy>=1.21.0,<2.0
pandas>=1.3.0
tqdm>=4.62.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
wandb>=0.12.0
tensorboard>=2.9.0
lmdb>=1.3.0
pytorch-lightning>=2.0.0
transformers
biopython>=1.79
faiss-gpu>=1.7.0
scipy>=1.7.0
seaborn>=0.11.0
PyYAML>=6.0 
matplotlib>=3.4.3
lmdb>=1.3.0
pytest>=7.0.0